Support HTTP POST requests (#44)
* allow caching of other HTTP methods like POST

* add contentlength and bodyhash key template vars

For caching POST requests it's important to distinguish requests between different body contents. This commit adds `http.request.contentlength` and `http.request.bodyhash`. For the body hash it's important that the cache key is calculated before the body has been read (before it's empty). Therefore the key is passed to `fetchUpstream`.